requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
aniso8601==0.82
gunicorn==19.0.0
itsdangerous==0.24
pytz==2014.4
six==1.7.2
Pillow==3.0.0
requests==2.8.1
imutils==0.3.1
